refactor(services): introduce RecipeImageService + move setPrimaryImage
Adds src/js/services/recipes/recipe-image-service.js as a sibling to
RecipeService. RecipeService keeps recipe-doc CRUD; RecipeImageService
will own per-image operations (lifecycle on the Storage files +
patches to entries inside recipes/{id}.images[]).

Migrations in this commit:

- setPrimaryImage moves from RecipeService to RecipeImageService.
  Same delegation under the hood (recipe-image-utils.setPrimaryImage);
  only the public entry-point class changed.
- image-approval-multi.js — the only setPrimaryImage caller — swaps
  its 4 call sites to RecipeImageService.setPrimaryImage. The
  RecipeService import in that file is replaced (was only used for
  setPrimaryImage).
- Tests: setPrimaryImage test moves out of recipe-service.test.mjs
  into recipe-image-service.test.mjs. Total test count unchanged
  (486 passing).

PR-H/1 of the AI-image-enhance work. Subsequent commits add
replaceImage (the AI-enhance verb) and migrate the
ai-image-enhancer/* modules.

Part of the service-layer refactor umbrella (#211).
